% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hfp.R
\name{clean_hfp_sf}
\alias{clean_hfp_sf}
\title{Clean HFP data files}
\usage{
clean_hfp_sf(hfp_temp, stop_times = TRUE)
}
\arguments{
\item{hfp_temp}{DATAFRAME that contains HFP 2.0 data}

\item{stop_times}{BOOLEAN true if stop times are kept in data}
}
\value{
cleaned dataframe
}
\description{
Removes outliers and selects variables needed in analysis
}
