% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_id_to_hfp.R
\name{hfp_sf_operations}
\alias{hfp_sf_operations}
\title{Attach link id to HFP rows}
\usage{
hfp_sf_operations(hfp_temp, links_temp, links_routes_temp, oday_temp, dir_temp)
}
\arguments{
\item{hfp_temp}{HFP dataframe}

\item{links_temp}{SPATIAL DATAFRAME which contains links and link_id field}

\item{oday_temp}{operating day}

\item{dir_temp}{direction}
}
\value{
HFP dataframe
}
\description{
Attach link id to HFP rows
}
