% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_id_to_hfp.R
\name{hfp_to_links_run}
\alias{hfp_to_links_run}
\title{Run points to links operations}
\usage{
hfp_to_links_run(hfp_temp, links_temp, links_routes_temp)
}
\arguments{
\item{hfp_temp}{DATAFRAME which contains cleaned HFP data file}

\item{links_temp}{SPATIAL DATAFRAME which contains links and link_id field}

\item{links_routes_temp}{DATAFRAME which has fields route_id, link_id, dir_id and stop_seq}
}
\value{
HFP dataframe that has link_id attached as new column
}
\description{
Run points to links operations
}
