% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_stats.R
\name{link_stats}
\alias{link_stats}
\title{Results from tidy HFP data}
\usage{
link_stats(
  data_path = "",
  tidy_path = "",
  result_path = "",
  start_time = 6,
  end_time = 19,
  tp_name = "vrk",
  links_shp = "links.shp",
  result_filename = "links_times.shp"
)
}
\arguments{
\item{data_path}{pathname of original data files, "" by default (graphical interface)}

\item{tidy_path}{pathname of tidy HFP files, "" by default (graphical interface)}

\item{result_path}{pathname where to write result shapefile, "" by default (graphical interface)}

\item{start_time}{starting hour, 6 by default}

\item{end_time}{ending hour (for example 16 => 16:59 last accepted), 19 by default}

\item{tp_name}{character suffix to result file variables, "vrk" by default}

\item{links_shp}{filename of links data in data_path, "links.shp" by default}

\item{result_filename}{filename of result shapefile in result_path, "links_times.shp" by default}
}
\description{
Result shapefile from tidy HFP data
}
\examples{
link_stats()
link_stats(start_time = 7, end_time = 9, result_filename = "result.shp")

JOIN TO SPATIAL FILE

}
