% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_hfp.R
\name{run_all_hfp}
\alias{run_all_hfp}
\title{Shapefile results from raw HFP data}
\usage{
run_all_hfp(
  data_path = "",
  tidy_path = "",
  result_path = "",
  routes_to_keep = "",
  start_time = 6,
  end_time = 19,
  tp_name = "vrk",
  links_shp = "links.shp",
  result_filename = "links_times.shp",
  links_routes = "links_routes.csv"
)
}
\arguments{
\item{data_path}{pathname of original data files, "" by default (graphical interface)}

\item{tidy_path}{pathname where to write tidy HFP files, "" by default (graphical interface)}

\item{result_path}{pathname where to write result shapefile, "" by default (graphical interface)}

\item{routes_to_keep}{character vector of selected routes, "" by default}

\item{start_time}{starting hour, 6 by default}

\item{end_time}{ending hour (for example 16 => 16:59 last accepted), 19 by default}

\item{tp_name}{character suffix to result file variables, "vrk" by default}

\item{links_shp}{filename of links data in data_path, "links.shp" by default}

\item{result_filename}{filename of result shapefile in result_path, "links_times.shp" by default}

\item{links_routes}{filename of route data in data_path, "links_routes.csv" by default}
}
\description{
User interface for using functions tidy_hfp() and link_stats()
}
\examples{
run_all_hfp()
run_all_hfp(routes_to_keep = c("1001", "1002"), start_time = 7, end_time = 9, links_shp = "example.shp")

}
