% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hfp.R
\name{tidy_hfp}
\alias{tidy_hfp}
\title{Tidy HFP data files from raw data}
\usage{
tidy_hfp(
  data_path = "",
  tidy_path = "",
  routes_to_keep = "",
  links_shp = "links.shp",
  links_routes = "links_routes.csv"
)
}
\arguments{
\item{data_path}{pathname of original data files, "" by default (graphical interface)}

\item{tidy_path}{pathname where to write tidy HFP files, "" by default (graphical interface)}

\item{routes_to_keep}{character vector of selected routes, "" by default}

\item{links_shp}{filename of links data in data_path, "links.shp" by default}

\item{links_routes}{filename of route data in data_path, "links_routes.csv" by default}
}
\description{
User interface for using functions clean_hfp_sf() and hfp_to_links_run()
}
\examples{
tidy_hfp()
tidy_hfp(routes_to_keep = c("1001", "1002"), links_shp = "example.shp")

}
